move index.js to src folder
